Revert "Update external JWT identity matching (#9769)"
This reverts commit 13e68be23caf8e82ecd25274b625293a16866240.
