+++ packages/definitions/src/fhir/r4/fhir.schema.json
@@ -62889,28 +62889,6 @@
     "IdentityProvider": {
       "description": "External Identity Provider (IdP) configuration details.",
       "properties": {
-        "issuer": {
-          "description": "Issuer URL for the external Identity Provider.",
-          "$ref": "#/definitions/string"
-        },
-        "audience": {
-          "description": "Audience URL for the external Identity Provider.",
-          "items": {
-            "$ref": "#/definitions/string"
-          },
-          "type": "array"
-        },
-        "tokenVerificationMethod": {
-          "description": "Token verification method for the external Identity Provider.",
-          "enum": [
-            "userinfo",
-            "jwks"
-          ]
-        },
-        "jwksUrl": {
-          "description": "Remote URL for the external Identity Provider JWKS endpoint.",
-          "$ref": "#/definitions/string"
-        },
         "authorizeUrl": {
           "description": "Remote URL for the external Identity Provider authorize endpoint.",
           "$ref": "#/definitions/string"
@@ -62975,7 +62953,14 @@
           ]
         }
       },
-      "additionalProperties": false
+      "additionalProperties": false,
+      "required": [
+        "authorizeUrl",
+        "tokenUrl",
+        "userInfoUrl",
+        "clientId",
+        "clientSecret"
+      ]
     },
     "ProjectMembership_Access": {
       "description": "Extended access configuration using parameterized access policies.",
