+++ packages/server/src/auth/external.ts
@@ -221,9 +221,6 @@ async function resolveExternalLoginIdentity(
     if (!email) {
       throw new OperationOutcomeError(badRequest('External token does not contain email address'));
     }
-    if (userInfo.email_verified !== true) {
-      throw new OperationOutcomeError(badRequest('External token email is not verified'));
-    }
     return { email };
   }
 
@@ -263,10 +260,6 @@ async function verifyExternalCode(
   code: string,
   codeVerifier: string | undefined
 ): Promise<Record<string, unknown>> {
-  if (!idp.tokenUrl) {
-    throw new OperationOutcomeError(badRequest('Missing token URL for external identity provider'));
-  }
-
   const headers: HeadersInit = {
     Accept: ContentType.JSON,
     'Accept-Encoding': 'identity',
@@ -283,10 +276,6 @@ async function verifyExternalCode(
   }
 
   if (idp.tokenAuthMethod === OAuthTokenAuthMethod.ClientSecretPost) {
-    if (!idp.clientId || !idp.clientSecret) {
-      throw new OperationOutcomeError(badRequest('Missing client ID or client secret for external identity provider'));
-    }
-
     params.append('client_id', idp.clientId);
     params.append('client_secret', idp.clientSecret);
   } else {
