+++ packages/server/src/auth/method.ts
@@ -47,7 +47,7 @@ export async function isExternalAuth(email: string): Promise<{ domain: string; a
   }
 
   const idp = domainConfig.identityProvider;
-  if (!idp?.authorizeUrl || !idp.clientId) {
+  if (!idp) {
     return undefined;
   }
