+++ packages/server/src/oauth/external.test.ts
@@ -5,13 +5,11 @@ import { encodeBase64Url, getReferenceString } from '@medplum/core';
 import type { Practitioner, Project, ProjectMembership } from '@medplum/fhirtypes';
 import { randomUUID } from 'crypto';
 import express from 'express';
-import { exportJWK, generateKeyPair, SignJWT } from 'jose';
 import request from 'supertest';
 import { vi } from 'vitest';
 import { inviteUser } from '../admin/invite';
 import { initApp, shutdownApp } from '../app';
-import { getConfig, loadTestConfig } from '../config/loader';
-import type { MedplumExternalAuthConfig } from '../config/types';
+import { loadTestConfig } from '../config/loader';
 import type { SystemRepository } from '../fhir/repo';
 import { getProjectSystemRepo } from '../fhir/repo';
 import { createTestProject } from '../test.setup';
@@ -24,7 +22,6 @@ describe('External auth', () => {
   const app = express();
   const npi = randomUUID();
   const externalSub = randomUUID();
-  const email = `external-${randomUUID()}@example.com`;
   let testProject: WithId<Project>;
   let practitioner: WithId<ProfileResource>;
   let systemRepo: SystemRepository;
@@ -67,15 +64,6 @@ describe('External auth', () => {
       firstName: 'External',
       lastName: 'User',
       externalId: externalSub,
-    });
-
-    await inviteUser({
-      project,
-      resourceType: 'Practitioner',
-      firstName: 'Email',
-      lastName: 'User',
-      email,
-      sendEmail: false,
     });
   });
 
@@ -228,243 +216,6 @@ describe('External auth', () => {
     expect(res.status).toBe(200);
   });
 
-  test('Success by JWKS verification', async () => {
-    const keyPair = await generateKeyPair('ES256');
-    const publicJwk = await exportJWK(keyPair.publicKey);
-    const jwksUrl = 'https://external-auth.example.com/.well-known/jwks.json';
-
-    await withExternalAuthProviders(
-      [
-        {
-          issuer: 'https://external-auth.example.com',
-          identityProvider: {
-            issuer: 'https://external-auth.example.com',
-            audience: ['external-api'],
-            jwksUrl,
-            tokenVerificationMethod: 'jwks',
-          },
-        },
-      ],
-      async () => {
-        fetchMock.mockImplementationOnce(() => mockFetchJson({ keys: [publicJwk] }));
-
-        const jwt = await new SignJWT({
-          nonce: randomUUID(),
-        })
-          .setProtectedHeader({ alg: 'ES256' })
-          .setIssuer('https://external-auth.example.com')
-          .setSubject(externalSub)
-          .setAudience('external-api')
-          .setIssuedAt()
-          .setExpirationTime('2h')
-          .sign(keyPair.privateKey);
-
-        const res = await request(app)
-          .get(`/oauth2/userinfo`)
-          .set('Authorization', 'Bearer ' + jwt);
-        expect(res.status).toBe(200);
-        expect(fetchMock).toHaveBeenCalledWith(jwksUrl, expect.anything());
-      }
-    );
-  });
-
-  test.each([
-    {
-      name: 'string audience match succeeds',
-      expectedAudience: ['external-api'],
-      actualAudience: 'external-api',
-      status: 200,
-    },
-    {
-      name: 'array audience match succeeds',
-      expectedAudience: ['external-api', 'alternate-api'],
-      actualAudience: ['wrong-api', 'alternate-api'],
-      status: 200,
-    },
-    {
-      name: 'audience mismatch is rejected',
-      expectedAudience: ['external-api'],
-      actualAudience: 'wrong-api',
-      status: 401,
-    },
-    {
-      name: 'missing audience is rejected when expected audience is configured',
-      expectedAudience: ['external-api'],
-      actualAudience: undefined,
-      status: 401,
-    },
-  ])('$name', async ({ expectedAudience, actualAudience, status }) => {
-    await withExternalAuthProviders(
-      [
-        {
-          issuer: 'https://external-auth.example.com',
-          identityProvider: {
-            audience: expectedAudience,
-            userInfoUrl: 'https://external-auth.example.com/oauth2/userinfo',
-          },
-        },
-      ],
-      async () => {
-        if (status === 200) {
-          fetchMock.mockImplementationOnce(() => mockFetchJson({ ok: true }));
-        }
-
-        const jwt = createFakeJwt({
-          iss: 'https://external-auth.example.com',
-          aud: actualAudience,
-          sub: externalSub,
-          nonce: randomUUID(),
-        });
-        const res = await request(app)
-          .get(`/oauth2/userinfo`)
-          .set('Authorization', 'Bearer ' + jwt);
-        expect(res.status).toBe(status);
-      }
-    );
-  });
-
-  test.each([
-    {
-      name: 'email to user membership',
-      identitySource: 'email' as const,
-      identityMappingMode: 'user-email' as const,
-      userInfo: () => ({ email, email_verified: true }),
-    },
-    {
-      name: 'subject to membership externalId',
-      identitySource: 'subject' as const,
-      identityMappingMode: 'project-membership-external-id' as const,
-      userInfo: () => ({ sub: externalSub }),
-    },
-    {
-      name: 'fhirUser to profile membership',
-      identitySource: 'fhir-user' as const,
-      identityMappingMode: 'project-membership-profile' as const,
-      userInfo: () => ({ fhirUser: getReferenceString(practitioner) }),
-    },
-  ])('Identity provider config maps $name', async ({ identitySource, identityMappingMode, userInfo }) => {
-    await withExternalAuthProviders(
-      [
-        {
-          issuer: 'https://external-auth.example.com',
-          identityProvider: {
-            authorizeUrl: 'https://external-auth.example.com/oauth2/authorize',
-            tokenUrl: 'https://external-auth.example.com/oauth2/token',
-            userInfoUrl: 'https://external-auth.example.com/oauth2/userinfo',
-            clientId: 'external-client',
-            clientSecret: 'external-secret',
-            identitySource,
-            identityMappingMode,
-          },
-        },
-      ],
-      async () => {
-        fetchMock.mockImplementationOnce(() => mockFetchJson(userInfo()));
-
-        const jwt = createFakeJwt({
-          iss: 'https://external-auth.example.com',
-          nonce: randomUUID(),
-        });
-        const res = await request(app)
-          .get(`/oauth2/userinfo`)
-          .set('Authorization', 'Bearer ' + jwt);
-        expect(res.status).toBe(200);
-      }
-    );
-  });
-
-  test('Identity provider legacy useSubject maps verified subject', async () => {
-    await withExternalAuthProviders(
-      [
-        {
-          issuer: 'https://external-auth.example.com',
-          identityProvider: {
-            userInfoUrl: 'https://external-auth.example.com/oauth2/userinfo',
-            useSubject: true,
-          },
-        },
-      ],
-      async () => {
-        fetchMock.mockImplementationOnce(() => mockFetchJson({ sub: externalSub }));
-
-        const jwt = createFakeJwt({
-          iss: 'https://external-auth.example.com',
-          nonce: randomUUID(),
-        });
-        const res = await request(app)
-          .get(`/oauth2/userinfo`)
-          .set('Authorization', 'Bearer ' + jwt);
-        expect(res.status).toBe(200);
-      }
-    );
-  });
-
-  test.each([
-    {
-      name: 'unverified email',
-      userInfo: { email, email_verified: false },
-    },
-    {
-      name: 'missing email_verified',
-      userInfo: { email },
-    },
-  ])('Identity provider email mapping rejects $name', async ({ userInfo }) => {
-    await withExternalAuthProviders(
-      [
-        {
-          issuer: 'https://external-auth.example.com',
-          identityProvider: {
-            userInfoUrl: 'https://external-auth.example.com/oauth2/userinfo',
-            identitySource: 'email',
-            identityMappingMode: 'user-email',
-          },
-        },
-      ],
-      async () => {
-        fetchMock.mockImplementationOnce(() => mockFetchJson(userInfo));
-
-        const jwt = createFakeJwt({
-          iss: 'https://external-auth.example.com',
-          nonce: randomUUID(),
-        });
-        const res = await request(app)
-          .get(`/oauth2/userinfo`)
-          .set('Authorization', 'Bearer ' + jwt);
-        expect(res.status).toBe(401);
-      }
-    );
-  });
-
-  test('Identity provider without bearer mapping uses JWT claims', async () => {
-    await withExternalAuthProviders(
-      [
-        {
-          issuer: 'https://external-auth.example.com',
-          identityProvider: {
-            authorizeUrl: 'https://external-auth.example.com/oauth2/authorize',
-            tokenUrl: 'https://external-auth.example.com/oauth2/token',
-            userInfoUrl: 'https://external-auth.example.com/oauth2/userinfo',
-            clientId: 'external-client',
-            clientSecret: 'external-secret',
-          },
-        },
-      ],
-      async () => {
-        fetchMock.mockImplementationOnce(() => mockFetchJson({ email: `wrong-${email}` }));
-
-        const jwt = createFakeJwt({
-          iss: 'https://external-auth.example.com',
-          fhirUser: getReferenceString(practitioner),
-          nonce: randomUUID(),
-        });
-        const res = await request(app)
-          .get(`/oauth2/userinfo`)
-          .set('Authorization', 'Bearer ' + jwt);
-        expect(res.status).toBe(200);
-      }
-    );
-  });
-
   test('Sub claim with caching', async () => {
     fetchMock.mockImplementationOnce(() => mockFetchJson({ ok: true }));
 
@@ -482,21 +233,6 @@ describe('External auth', () => {
       .get(`/oauth2/userinfo`)
       .set('Authorization', 'Bearer ' + jwt);
     expect(res2.status).toBe(200);
-  });
-
-  test('Expired token is rejected before cache lookup', async () => {
-    fetchMock.mockClear();
-
-    const jwt = createFakeJwt({
-      iss: 'https://external-auth.example.com',
-      sub: externalSub,
-      exp: Math.floor(Date.now() / 1000) - 1,
-    });
-    const res = await request(app)
-      .get(`/oauth2/userinfo`)
-      .set('Authorization', 'Bearer ' + jwt);
-    expect(res.status).toBe(401);
-    expect(fetchMock).not.toHaveBeenCalled();
   });
 
   test('Sub claim with unknown externalId', async () => {
@@ -605,18 +341,4 @@ describe('External auth', () => {
 
 function createFakeJwt(claims: Record<string, unknown>): string {
   return `header.${encodeBase64Url(JSON.stringify(claims))}.signature`;
-}
-
-async function withExternalAuthProviders(
-  externalAuthProviders: MedplumExternalAuthConfig[],
-  fn: () => Promise<void>
-): Promise<void> {
-  const savedExternalAuthProviders = getConfig().externalAuthProviders;
-  getConfig().externalAuthProviders = externalAuthProviders;
-
-  try {
-    await fn();
-  } finally {
-    getConfig().externalAuthProviders = savedExternalAuthProviders;
-  }
 }
